In all the methods used for reading or writing memory, use uint32_t for the size arguments instead of size_t. 1. We won't be reading and writing gigabytes or more. 2. I don't want the size of the type to change based on the instruction set.

+++ DebugEngine/Exec/CommandFunctor.h
@@ -121,9 +121,9 @@ namespace MagoCore
         IProcess*       Process;
         Address         Address;
         uint8_t*        Buffer;
-        SIZE_T          Length;
-        SIZE_T          OutLengthRead;
-        SIZE_T          OutLengthUnreadable;
+        uint32_t        Length;
+        uint32_t        OutLengthRead;
+        uint32_t        OutLengthUnreadable;
 
         ReadMemoryParams( Exec& exec )
             :   ExecCommandFunctor( exec ),
@@ -145,8 +145,8 @@ namespace MagoCore
         IProcess*       Process;
         Address         Address;
         uint8_t*        Buffer;
-        SIZE_T          Length;
-        SIZE_T          OutLengthWritten;
+        uint32_t        Length;
+        uint32_t        OutLengthWritten;
 
         WriteMemoryParams( Exec& exec )
             :   ExecCommandFunctor( exec ),
