+++ DebugEngine/MagoNatDE/DebuggerProxy.h
@@ -44,16 +44,16 @@ namespace Mago
         HRESULT ReadMemory( 
             ICoreProcess* process, 
             Address address,
-            SIZE_T length, 
-            SIZE_T& lengthRead, 
-            SIZE_T& lengthUnreadable, 
+            uint32_t length, 
+            uint32_t& lengthRead, 
+            uint32_t& lengthUnreadable, 
             uint8_t* buffer );
 
         HRESULT WriteMemory( 
             ICoreProcess* process, 
             Address address,
-            SIZE_T length, 
-            SIZE_T& lengthWritten, 
+            uint32_t length, 
+            uint32_t& lengthWritten, 
             uint8_t* buffer );
 
         HRESULT SetBreakpoint( ICoreProcess* process, Address address );
